+++ plutus-numeric/test/golden/Main.hs
@@ -3,8 +3,10 @@ module Main (main) where
 import PlutusTx.NatRatio (NatRatio)
 import PlutusTx.Natural (Natural)
 import PlutusTx.Positive (Positive)
-import PlutusTx.Rational (Rational)
+import PlutusTx.Ratio (Rational)
 import Test.Tasty (defaultMain, testGroup)
+import Test.QuickCheck.Plutus.Instances () 
+import PlutusTx.Numeric.Orphans ()
 import Test.Tasty.Plutus.Golden (goldenData, goldenJSON, goldenToSchema)
 import Prelude hiding (Rational)